Add a queue-disabled mode.
Many systems already store their pending/outgoing message queue on persistent
storage (e.g. in a database). If transmission of a particular message fails,
these systems don't need an in-memory queue and/or an automatic retry; the next
database query will pull out all of the unsent messages, and they'll be
submitted to `send` again, regardless of whether they've had prior attempts.

If you're using node-gammu-json with such a system, pass `{ queue: false }` to
the constructor's `_options` argument, and node-gammu-json will not
automatically reattempt transmission of a failed `send` operation.